# backend/api.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: "frigo"
  description: ""
  license:
     name: GPLv3
servers:
   - url: http://localhost:8000/api/v1

paths:
  /fridge:
    get:
      description: Get all the content of one fridge
      parameters:
        - name: ownedBy
          in: query
          description: Filters items by owner
          required: false
          schema:
            type: string
        - name: expiresIn
          in: query
          description: Filters items expiring in a given number of days
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: 'Collection of items'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      description: 'Add an item to the fridge'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemPost"
      responses:
        '200':
          description: "New item added"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"

  /fridge/{itemId}:
    get:
      description: Get the item with the specified itemId
      responses:
        '200':
          description: 'Searched item'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
    patch:
      description: Set the item with the specified itemId as Opened
      responses:
        '200':
          description: 'Item updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
    delete:
      description: Delete the item with the specified itemId
      responses:
        '200':
          description: 'Item updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'

components:
  schemas:
    Item:
      type: object
      required:
        - id
        - name
        - ownedBy
        - categoryId
        - amount
        - unit
        - expDate
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLenght: 1
          maxLength: 30
        ownedBy:
          type: string
          minLenght: 1
          maxLength: 20
        categoryId:
          type: integer
        amount:
          type: number
          format: float
        unit:
          type: string
          enum:
            - g
            - kg
            - ml
            - l
            - piece
        expDate: 
          type: string
          format: date
          readOnly: true
        openedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          
    ItemPost:
      type: object
      required:
        - name
        - ownedBy
        - categoryId
      properties:
        name:
          type: string
          minLenght: 1
          maxLength: 30
        ownedBy:
          type: string
          minLenght: 1
          maxLength: 20
        categoryId:
          type: integer
        amount:
          type: number
          format: float
          nullable: true
        unit:
          type: string
          enum:
            - g
            - kg
            - ml
            - l
            - piece
          nullable: true
        expDate: 
          type: string
          format: date
          nullable: true

    Category:
      type: object
      required:
        - name
        - shelfLife
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 20
        shelfLife:
          type: integer
          description: Shelf life in days
        openedLife:
          type: integer
          nullable: true
          description: Shelf life in days after opening